QUASAR: Lasso select
Using jQuery selectable, one can now draw a box around the elements one 
wishes to move and then drag them as one; the algorithm for determining 
the new x and y positions for the selected elements aside from the one
being dragged needs refining a bit, as it maintains the same general shape 
but introduces a little inconsistency in the distances between the nodes.

In order to get this to work I had to end up modifying the jquery.ui
source file for the delectable module, as the helper element (in this case the
select lasso) was rendering incorrectly, though the logic in regards to the 
event were correct. This is a known bug in selectable (#4377), and as one of
the proposed fixes is along the lines of what I did to correct offset problems
with the draggables I added it myself. Until this problem is fixed in the main
jQuery Selectable CDN we'll have to reference the local copy.

#12