QUASAR: Call handleOffscreenIndicators after prepopulating nodeCanvas
Before, if you moved elements off the viewable pane of the canvas and 
reloaded the page, there would be nothing to show that those nodes were
off-canvas. Now, at the end of the prepop method, the function that 
handles the calculation and positioning of the off-canvas indicators, which
generates the offscreen indicators correctly.

#12