QUASAR: Final fix for $hoverInfo timing issues
Found that the .stop() method was the cure for hoverInfo timing concerns.
The problem was the queue for the animations was being executed in 
sequence, regardless of the need to stop a fadeIn to fadeOut, etc. .stop()
fixed this. There's still a problem with rapid mouseovers (for whatever
reason) aligning the left and top of the hoverInfo to the left and top of the
nodeCanvas; this is masked by the rapidness of the fadeIn/fadeOut with
.stop() implemented, but it shouldn't be happening regardless. Will 
investigate further.

#12