CERES: Use SharedPreferences to maintain persistent data
Except for password to Ceres. Going to need to use an AuthToken
based system for that, but for now everything else seems to work
pretty well. Going to go through and test it a little more, weird 
things happen when the app is force closed that need to get 
figured out.