CERES: Multiple new features
- Created a listener to watch for battery_low events; when these
are fired, the update thread for the grid stops, to preserve power
and reduce network activity from the app. The grid can still
be manually refreshed. 

- Created a listener for network connection; if you're not 
connected to the network, the grid won't try to update 
automatically.

- Added search bar to grid. Accepts IPs and interface names/
aliases. If you put in an IP with no *, the activity checks for exact
equality; if you put in, say, 192.168.*, it simply checks for contains
and returns that. Still have some edge cases to work out, but the
core is there.

- Made refresh a button instead of an item in the dropdown menu
on the grid page.