CERES: Move over to standard Java HTTPSUrlConnection
Makes it easier to have a self-signed certificate AND 
hostname verification; a problem with the previous 
implementation was the lack of the latter. In order to use a 
self-signed certificate, simply generate a .crt file (hard-coded
"certificate.crt" for now) and place it into /res/raw/, recompile
the app, and load it onto the phone. 

close #791