QUASAR: Fix some compat problems with new jquery ui stuff
There's a bug report up right now for jquery ui related to 
the selectable lasso; in the vanilla implementation, it won't be
drawn to the right place on the page, because it doesn't take into
account the fact that the appendTo element may have it's 
own offsets to take into account. To skirt around this, changed 
the source, and it works. Had to minify it in a pretty ghetto way,
thus the change to that file, but at least it's more compact than
the normal file. Also had to deal with some styling stuff and 
problems with resizing the canvas.