Quasar: Shift from hoverInfo to slideInfo
A div slides out from the right that has the same info as the 
canvasElement hovers, but not plagued with the same amount
of problems, from what I've seen. It's responsive, it's quick, it
looks good. I'm keepin it. Still need to deal with the case where
a selected canvasElement is deleted (the slideInfo div doesn't 
retract) but that's a small thing.

#12