QUASAR: Some fixes for hide/show flip-flopping on elements that are close
If two of the canvasElements are close to each other (close enough that
mouseleave for one is mouseenter/mouseover for the other), the 
original 'show' for slideInfo would trigger, the 'hide' for the same would
trigger on the mouseleave, but the 'show' for the second element wouldn't.
Fixed that, it'll just stay up until the cursor is on the canvas.

#12