QUASAR: If a canvasElement is selected and deleted, hide slideInfo
Fixing a quirk referenced in one of my last couple of commits; basically,
before this commit when one of the canvas Node elements is clicked on/
lassoed, the slideInfo pane stays open, and able to be modified. If said
node is then deleted, either with the delete key or by the context menu, 
the slideInfo menu would stay open, reflecting stale information. Doesn't
happen anymore. 

#12