QUASAR: Mitigate a little css problem with slideInfo
When the show()/hide() functions are used, jQueryUI adds 
a class 'ui-effects-wrapper' that is used, it seems to be an anchor
for user-made CSS to be applied to the sliding element. Had to
add border-radius to help the right side of the sliding element not
be so sharp, though it still does sometime. I don't think it's a 
completely solvable problem, but the duration for the show()/
hide() animation is so short that you probably wouldn't
notice it.

#12