QUASAR: Fix bug for applying containment to canvasElements
Noticed that, if a node is created whose count is greater than 1, 
the drag/drop functionality would error out after the topology
was saved and the page refreshed. This was because of some 
holdover code and a reference to the wrong attribute of an object
when the handlers for drag/drop, etc. were being applied. Seems
fixed now.

#12