QUASAR: Allow for any-length profileInfo fadeIns
Tried messing around with jScrollPane but it wasn't having it, so just made
the profileInfo fadeIn bubble have a fixed max-height and scroll after that.
It doesn't look super pretty, but it's the best that can be done given the 
inability to use styling for scrollbars in a reasonable, cross-platform way.

#12