QUASAR: Move Wysiwyg back into its own page
Dealing with the display concurrency for all the events of 
haystack config manage are a little too overwhelming for the 
timeframe, so I moved WYSIWYG into a button on the Nodes
page. Need to check for profile/mac mismatch for nodes
when using a Topo file, because there's a possibility profiles can
change/nodes added or deleted on other pages that aren't 
reflected in the topology file.

#12