QUASAR: Manage synchronization issues between topo file and real nodes
If a user removed a node/profile from a haystack, and had a topo file from
an earlier wysiwyg interaction that had obsolete nodes, the wysiwyg would
read from the topology file and present nodes on the canvas that aren't
correct. Fixed that, adding nodes from the topology file now checks for 
node existence both from a mac perspective and from their profile. 

#12