QUASAR: Rewrite help box on Wysiwyg
The instructions that were left there before were intended as a 
style placeholder for future revision, but with the moving around
of the wysiwyg page the styles got messed up and the hierarchy 
looked bad. Fixed it, reworded the instructions to (hopefully)
be more clear. May need to add more info relating to things like
the tabs.

#12